Move ChannelContextType to graphql/channel module
