+++ saleor/graphql/product/types/products.py
@@ -3,8 +3,6 @@ from dataclasses import asdict
 import graphene
 from django.conf import settings
 from graphene import relay
-from graphene.types.resolver import get_default_resolver
-from graphene_django.types import DjangoObjectType
 from graphene_federation import key
 from graphql.error import GraphQLError
 
@@ -28,6 +26,7 @@ from ....warehouse.availability import (
 )
 from ...account.enums import CountryCodeEnum
 from ...channel import ChannelContext, ChannelQsContext
+from ...channel.types import ChannelContextType
 from ...channel.utils import get_default_channel_or_graphql_error
 from ...core.connection import CountableDjangoObjectType
 from ...core.enums import ReportingPeriod, TaxRateType
@@ -42,7 +41,6 @@ from ...discount.dataloaders import DiscountsByDateTimeLoader
 from ...meta.deprecated.resolvers import resolve_meta, resolve_private_meta
 from ...meta.types import ObjectWithMetadata
 from ...translations.fields import TranslationField
-from ...translations.resolvers import resolve_translation
 from ...translations.types import (
     CategoryTranslation,
     CollectionTranslation,
@@ -125,43 +123,6 @@ class ProductPricingInfo(BasePricingInfo):
 
     class Meta:
         description = "Represents availability of a product in the storefront."
-
-
-class ChannelContextType(DjangoObjectType):
-    """A Graphene type that supports resolvers' root as ChannelContext objects."""
-
-    class Meta:
-        abstract = True
-
-    @staticmethod
-    def resolver_with_context(
-        attname, default_value, root: ChannelContext, info, **args
-    ):
-        resolver = get_default_resolver()
-        return resolver(attname, default_value, root.node, info, **args)
-
-    @staticmethod
-    def resolve_id(root: ChannelContext, _info):
-        return root.node.pk
-
-    @classmethod
-    def is_type_of(cls, root: ChannelContext, info):
-        return super().is_type_of(root.node, info)
-
-    @staticmethod
-    def resolve_metadata(root: ChannelContext, info):
-        # Used in metadata API to resolve metadata fields from an instance.
-        return ObjectWithMetadata.resolve_metadata(root.node, info)
-
-    @staticmethod
-    def resolve_private_metadata(root: ChannelContext, info):
-        # Used in metadata API to resolve private metadata fields from an instance.
-        return ObjectWithMetadata.resolve_private_metadata(root.node, info)
-
-    @staticmethod
-    def resolve_translation(root: ChannelContext, info, language_code):
-        # Resolver for TranslationField; needs to be manually specified.
-        return resolve_translation(root.node, info, language_code)
 
 
 @key(fields="id")
